requirements.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
transformers==4.28.1
torch==2.0.0
streamlit==1.22.0
plotly==5.14.1
python-dotenv==1.0.0
PyYAML==6.0
textblob==0.17.1
pytz==2023.3
vaderSentiment==3.3.2 